+++ src/Catalogo/ArticuloCardCatalogo.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import { ArticuloVentaDTO } from "../ts/Clases";
+import { ArticuloVentaDTO } from "../../ts/Clases";
 import DetalleArticulo from "./DetalleArticulo";
-import { obtenerImagen } from "../ts/Imagen";
+import { obtenerImagen } from "../../ts/Imagen";
 
 interface Props {
     articulo: ArticuloVentaDTO;
@@ -21,7 +21,7 @@ export default function ArticuloCardCatalogo({articulo}: Props) {
                 <div className="bg-[#99999959] w-full rounded-2xl flex flex-col gap-1">
                     {/**Imagen y tiempo */}
                     <div className="relative">
-                        <img className=" w-10/12 mt-2 mb-0 m-auto rounded-2xl" src={obtenerImagen(articulo.imagen)} alt="No se encontro la imagen" />
+                        <img className=" w-10/12 mt-2 mb-0 m-auto rounded-2xl" src={obtenerImagen(articulo.imagenArticulo)} alt="No se encontro la imagen" />
                         {articulo.tiempoEstimado && (
                             <div className="absolute bottom-0 left-1/12 bg-white m-auto text-center text-black p-1 rounded-bl-2xl rounded-tr-2xl ">
                                 <h1 className="text-xl">{`${articulo.tiempoEstimado.split(" ")[0]}-${Number(articulo.tiempoEstimado.split(" ")[0]) + 5}`} min</h1>
